+++ src/gantt.ts
@@ -33,6 +33,7 @@ module powerbi.extensibility.visual {
     import DataView = powerbi.DataView;
     import IViewport = powerbi.IViewport;
     import VisualObjectInstance = powerbi.VisualObjectInstance;
+    import DataViewCategoryColumn = powerbi.DataViewCategoryColumn;
     import VisualObjectInstanceEnumeration = powerbi.VisualObjectInstanceEnumeration;
     import DataViewObjectPropertyIdentifier = powerbi.DataViewObjectPropertyIdentifier;
     import EnumerateVisualObjectInstancesOptions = powerbi.EnumerateVisualObjectInstancesOptions;
@@ -123,14 +124,14 @@ module powerbi.extensibility.visual {
         "second",
         "minute",
         "hour",
-        "day"
+        "day",
     ];
 
     export enum DurationUnits {
         Second = <any>"second",
         Minute = <any>"minute",
         Hour = <any>"hour",
-        Day = <any>"day"
+        Day = <any>"day",
     }
 
      export enum DateTypes {
@@ -200,16 +201,19 @@ module powerbi.extensibility.visual {
         tasks: Task[];
         legendData: LegendData;
         taskTypes: TaskTypes;
+        isDurationFilled: boolean;
     }
 
     export interface TaskTypes { /*TODO: change to more proper name*/
         typeName: string;
-        types: TaskTypesList;
+        types: TaskTypeMetadata[];
     }
 
-    export type TaskTypesList = {
-        [group: string]: string[]
-    } | undefined[];
+    export interface TaskTypeMetadata {
+        name: string;
+        columnGroup: DataViewValueColumnGroup;
+        selectionColumn: DataViewCategoryColumn;
+    }
 
     export interface GanttCalculateScaleAndDomainOptions {
         viewport: IViewport;
@@ -527,16 +531,14 @@ module powerbi.extensibility.visual {
         */
         private static getTooltipInfo(
             task: Task,
+            locale: string,
             formatters: GanttChartFormatters,
             durationUnit: string): VisualTooltipDataItem[] {
 
             let tooltipDataArray: VisualTooltipDataItem[] = [];
             const durationLabel: string = Gantt.generateLabelForDuration(task.duration, durationUnit);
             if (task.taskType) {
-                tooltipDataArray.push({
-                    displayName: "Legend",
-                    value: task.taskType
-                });
+                tooltipDataArray.push({ displayName: "Legend", value: task.taskType });
             }
 
             tooltipDataArray.push({
@@ -562,10 +564,7 @@ module powerbi.extensibility.visual {
             });
 
             if (task.resource) {
-                tooltipDataArray.push({
-                    displayName: "Resource",
-                    value: task.resource
-                });
+                tooltipDataArray.push({ displayName: "Resource", value: task.resource });
             }
 
             for (const key of Object.keys(task.extraInformation)) {
@@ -640,7 +639,10 @@ module powerbi.extensibility.visual {
             host: IVisualHost,
             colorPalette: IColorPalette,
             settings: GanttSettings,
-            taskTypes: TaskTypes): LegendData {
+            taskTypes: TaskTypes,
+            useDefaultColor: boolean): LegendData {
+
+            const colorHelper = new ColorHelper(colorPalette, Gantt.LegendPropertyIdentifier);
             const legendData: LegendData = {
                 fontSize: settings.legend.fontSize,
                 dataPoints: [],
@@ -648,23 +650,26 @@ module powerbi.extensibility.visual {
                 labelColor: settings.legend.labelColor
             };
 
-            const colorHelper = new ColorHelper(colorPalette, Gantt.LegendPropertyIdentifier);
+            legendData.dataPoints = taskTypes.types.map(
+                (typeMeta: TaskTypeMetadata): LegendDataPoint => {
+                    let color: string = settings.taskConfig.fill;
 
-            _.each(taskTypes.types, (tasksNames, groupName) => {
-                let color: string = colorHelper.getColorForSeriesValue(undefined, groupName),
-                    selectionId: ISelectionId = host.createSelectionIdBuilder()
-                        .withMeasure("Legend." + groupName)
-                        .createSelectionId();
+                    if (!useDefaultColor) {
+                        color = (taskTypes.types.length <= 1)
+                            ? settings.taskConfig.fill
+                            : colorHelper.getColorForMeasure(typeMeta.columnGroup.objects, typeMeta.name);
+                    }
 
-                legendData.dataPoints.push({
-                    color: color,
-                    icon: LegendIcon.Circle,
-                    label: groupName,
-                    identity: selectionId,
-                    selected: false
+                    return {
+                        label: typeMeta.name as string,
+                        color: color,
+                        icon: LegendIcon.Circle,
+                        selected: false,
+                        identity: host.createSelectionIdBuilder()
+                            .withCategory(typeMeta.selectionColumn, 0)
+                            .createSelectionId()
+                    };
                 });
-
-            });
 
             return legendData;
         }
@@ -673,54 +678,76 @@ module powerbi.extensibility.visual {
         * Create task objects dataView
         * @param dataView The data Model.
         * @param formatters task attributes represented format.
+        * @param taskColor Color of task
         * @param settings settings of visual
+        * @param colors colors of groped tasks
         * @param host Host object
         * @param taskTypes
-        * @param legendData data for legend
         */
         private static createTasks(
             dataView: DataView,
             taskTypes: TaskTypes,
             host: IVisualHost,
             formatters: GanttChartFormatters,
+            colors: IColorPalette,
             settings: GanttSettings,
-            legendData: LegendData): Task[] {
+            taskColor: string): Task[] {
 
             const tasks: Task[] = [];
+            const colorHelper: ColorHelper = new ColorHelper(colors, Gantt.LegendPropertyIdentifier);
             const values: GanttColumns<any> = GanttColumns.getCategoricalValues(dataView);
+            const groupValues: GanttColumns<DataViewValueColumn>[] = GanttColumns.getGroupedValueColumns(dataView);
 
             if (!values.Task) {
                 return tasks;
             }
             values.Task.forEach((categoryValue: PrimitiveValue, index: number) => {
-                let durationUnit: string = settings.general.durationUnit || "day";
+                let duration: number = settings.general.durationMin;
+                let durationUnit: string = settings.general.durationUnit;
+                let color: string = taskColor || Gantt.DefaultValues.TaskColor;
+                let completion: number = 0;
+                let taskType: TaskTypeMetadata = null;
                 let wasDowngradeDurationUnit: boolean = false;
 
                 const selectionBuider: ISelectionIdBuilder = host
                     .createSelectionIdBuilder()
                     .withCategory(dataView.categorical.categories[0], index);
 
-                let duration: number = ((values.Duration && values.Duration[index] > settings.general.durationMin)
-                        && values.Duration[index] as number) || settings.general.durationMin;
+                if (groupValues) {
+                    groupValues.forEach((group: GanttColumns<DataViewValueColumn>) => {
+                        if (group.Duration && group.Duration.values[index] !== null) {
+                            taskType = _.find(taskTypes.types,
+                                (typeMeta: TaskTypeMetadata) => typeMeta.name === group.Duration.source.groupName);
 
-                if (duration && duration % 1 !== 0) {
-                    durationUnit = Gantt.downgradeDurationUnit(durationUnit);
-                    let stepDurationTransformation: number =
-                        GanttDurationUnitType.indexOf(settings.general.durationUnit) - GanttDurationUnitType.indexOf(durationUnit);
+                            if (taskType) {
+                                selectionBuider.withCategory(taskType.selectionColumn, 0);
+                                color = colorHelper.getColorForMeasure(taskType.columnGroup.objects, taskType.name);
+                            }
 
-                    duration = Gantt.transformDuration(duration, durationUnit, stepDurationTransformation);
-                    wasDowngradeDurationUnit = true;
-                }
+                            duration = group.Duration.values[index] > settings.general.durationMin
+                                && group.Duration.values[index] as number;
 
-                let completion: number = ((values.Completion && values.Completion[index])
-                    && Gantt.convertToDecimal(values.Completion[index] as number)) || 0;
+                            if (duration && duration % 1 !== 0) {
+                                durationUnit = Gantt.downgradeDurationUnit(durationUnit);
+                                let stepDurationTransformation: number =
+                                    GanttDurationUnitType.indexOf(settings.general.durationUnit) - GanttDurationUnitType.indexOf(durationUnit);
 
-                if (completion < Gantt.ComplectionMin) {
-                    completion = Gantt.ComplectionMin;
-                }
+                                duration = Gantt.transformDuration(duration, durationUnit, stepDurationTransformation);
+                                wasDowngradeDurationUnit = true;
+                            }
 
-                if (completion > Gantt.ComplectionMax) {
-                    completion = Gantt.ComplectionMax;
+                            completion = ((group.Completion && group.Completion.values[index])
+                                && Gantt.convertToDecimal(group.Completion.values[index] as number)) || 0;
+
+                            if (completion < Gantt.ComplectionMin) {
+                                completion = Gantt.ComplectionMin;
+                            }
+
+                            if (completion > Gantt.ComplectionMax) {
+                                completion = Gantt.ComplectionMax;
+                            }
+                        }
+                    });
                 }
 
                 const selectionId: powerbi.extensibility.ISelectionId = selectionBuider.createSelectionId();
@@ -743,9 +770,6 @@ module powerbi.extensibility.visual {
                     }
                 }
 
-                let taskType: string = Gantt.getTaskType(taskTypes.types, <string>categoryValue);
-                let color: string = Gantt.getColorByTaskType(legendData.dataPoints, taskType, settings);
-
                 const task: Task = {
                     color,
                     completion,
@@ -755,7 +779,7 @@ module powerbi.extensibility.visual {
                     start: startDate,
                     end: null,
                     duration,
-                    taskType,
+                    taskType: taskType && taskType.name,
                     description: categoryValue as string,
                     tooltipInfo: [],
                     selected: false,
@@ -766,7 +790,7 @@ module powerbi.extensibility.visual {
                 };
 
                 task.end = d3.time[durationUnit].offset(task.start, task.duration);
-                if (settings.daysOff.show) {
+                if (settings.daysOff.show && duration) {
                     let datesDiff: number = 0;
                     do {
                         task.daysOffList = Gantt.calculateDaysOff(
@@ -790,41 +814,12 @@ module powerbi.extensibility.visual {
                     } while (task.daysOffList.length && datesDiff - DaysInAWeekend > DaysInAWeek);
                 }
 
-                task.tooltipInfo = Gantt.getTooltipInfo(task, formatters, durationUnit);
+                task.tooltipInfo = Gantt.getTooltipInfo(task, host.locale, formatters, durationUnit);
 
                 tasks.push(task);
             });
 
             return tasks;
-        }
-
-        private static getTaskType(
-            taskTypes: TaskTypesList,
-            categoryValue: string): string {
-            for (let taskType in taskTypes) {
-                if (taskTypes[taskType] &&
-                    taskTypes.hasOwnProperty(taskType) &&
-                    taskTypes[taskType].indexOf(categoryValue) !== -1) {
-                    return taskType;
-                }
-            }
-        }
-
-        private static getColorByTaskType(
-            dataPoints: any[],
-            taskType: string,
-            settings: GanttSettings): string {
-            let color: string = (dataPoints.length <= 1)
-                ? settings.taskConfig.fill
-                : Gantt.DefaultValues.TaskColor;
-
-            dataPoints.forEach(item => {
-                if (item.label === taskType) {
-                    color = item.color;
-                }
-            });
-
-            return color;
         }
 
         private static downgradeDurationUnit(durationUnit: string): string {
@@ -858,6 +853,7 @@ module powerbi.extensibility.visual {
                 default:
                     return duration;
             }
+
         }
 
         private static transformDuration(
@@ -1014,23 +1010,27 @@ module powerbi.extensibility.visual {
             const settings: GanttSettings = GanttSettings.parse<GanttSettings>(dataView);
             const taskTypes: TaskTypes = Gantt.getAllTasksTypes(dataView);
             const formatters: GanttChartFormatters = this.getFormatters(dataView, settings, host.locale || null);
-            const legendData = Gantt.createLegend(host, colors, settings, taskTypes);
 
-            const tasks: Task[] = Gantt.createTasks(
-                dataView,
-                taskTypes,
-                host,
-                formatters,
-                settings,
-                legendData
-            );
+            const index: number = _.findIndex(dataView.metadata.columns, col => col.roles.hasOwnProperty(GanttRoles.Duration));
+            let isDurationFilled: boolean = true;
 
+            if (index === -1) {
+                isDurationFilled = false;
+            }
+            const legendData = Gantt.createLegend(host, colors, settings, taskTypes, !isDurationFilled);
+
+            let taskColor: string = (legendData.dataPoints.length <= 1) || !isDurationFilled
+                ? settings.taskConfig.fill
+                : null;
+
+            const tasks: Task[] = Gantt.createTasks(dataView, taskTypes, host, formatters, colors, settings, taskColor);
             return {
                 dataView,
                 settings,
                 taskTypes,
                 tasks,
-                legendData
+                legendData,
+                isDurationFilled
             };
         }
 
@@ -1059,29 +1059,27 @@ module powerbi.extensibility.visual {
                 typeName: "",
                 types: []
             };
+            let index: number = _.findIndex(dataView.metadata.columns, col => col.roles.hasOwnProperty(GanttRoles.Legend));
 
-            let indexLegend: number = _.findIndex(dataView.categorical.categories,
-                col => col.source.roles.hasOwnProperty(GanttRoles.Legend));
-
-            if (indexLegend !== -1) {
-                const categories = dataView.categorical.categories;
-                let groupedValues: any = {};
-                let indexTask: number = _.findIndex(dataView.categorical.categories,
-                    col => col.source.roles.hasOwnProperty(GanttRoles.Task));
-
-                taskTypes.typeName = dataView.metadata.columns[indexLegend].displayName;
-
-                categories[indexTask].values.forEach((item, idx) => {
-                    let taskName: PrimitiveValue = <string>categories[indexTask].values[idx];
-                    let taskLegend: PrimitiveValue = <string>categories[indexLegend].values[idx];
-                    if (!groupedValues[taskLegend]) {
-                        groupedValues[taskLegend] = [];
-                    }
-
-                    groupedValues[taskLegend].push(taskName);
+            if (index !== -1) {
+                taskTypes.typeName = dataView.metadata.columns[index].displayName;
+                let legendMetaCategoryColumn: DataViewMetadataColumn = dataView.metadata.columns[index];
+                let groupValues = dataView.categorical.values.grouped();
+                taskTypes.types = groupValues.map((group: DataViewValueColumnGroup): TaskTypeMetadata => {
+                    let column: DataViewCategoryColumn = {
+                        identity: [group.identity],
+                        source: {
+                            displayName: null,
+                            queryName: legendMetaCategoryColumn.queryName
+                        },
+                        values: null
+                    };
+                    return {
+                        name: group.name as string,
+                        selectionColumn: column,
+                        columnGroup: group
+                    };
                 });
-
-                taskTypes.types = groupedValues;
             }
 
             return taskTypes;
@@ -1920,6 +1918,10 @@ module powerbi.extensibility.visual {
         }
 
         private enumerateLegend(instanceEnumeration: VisualObjectInstanceEnumeration): VisualObjectInstance[] {
+            if (!this.viewModel.isDurationFilled) {
+                return;
+            }
+
             const dataPoints: LegendDataPoint[] = this.viewModel && this.viewModel.legendData.dataPoints;
             if (!dataPoints || !(dataPoints.length > 0)) {
                 return;